pyproject.toml
[tool.black]
line-length = 120
target-version = ["py311"]
include = ".*py$"

[tool.isort]
profile = "black"
force_single_line = false

[tool.poetry]
name = "airy-rewrite"
version = "0.1.0"
description = ""
authors = ["CraazzzyyFoxx <andrey.pod19@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "airy_rewrite"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
hikari-lightbulb = "^2.3.1"
hikari-miru = "^2.0.1"
lru-dict = "^1.1.8"
orjson = "^3.8.3"
pydantic = "^1.10.2"
deepdiff = "^6.2.2"
python-dotenv = "^0.21.0"
dateparser = "^1.1.4"
click = "^8.1.3"
loguru = "^0.6.0"
aioredis = "^2.0.1"
asyncpg = "^0.27.0"
parsedatetime = "^2.6"
levenshtein = "^0.20.9"
fuzzywuzzy = "^0.18.0"
pygount = "^1.5.1"

[tool.poetry.dev-dependencies]



[tool.poetry.group.dev.dependencies]
mypy = "^0.991"
black = "^22.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
